Refactoring: Update G2P to use text2text-generation
