+++ docker_images/speechbrain/app/pipelines/text2text_generation.py
