+++ docker_images/speechbrain/tests/test_api.py
@@ -32,7 +32,7 @@ TESTABLE_MODELS: Dict[str, List[str]] = {
     "text-to-speech": [
         "speechbrain/tts-tacotron2-ljspeech",
     ],
-    "text-to-text": [
+    "text2text-generation": [
         # SoundChoice G2P
         "speechbrain/soundchoice-g2p"
     ]
